refactor: move legacy support to separate service
