+++ src/common/constants/ARC0027.ts
@@ -1,5 +1,3 @@
-export const ARC_0027_PREFIX: string = 'arc0027';
-export const ARC_0027_CHANNEL_NAME: string = `${ARC_0027_PREFIX}:channel`;
 export const ARC_0027_DEFAULT_REQUEST_TIMEOUT: number = 180000; // 3 minutes in milliseconds
 export const ARC_0027_LOWER_REQUEST_TIMEOUT: number = 60000; // 1 minute in milliseconds
 export const ARC_0027_UPPER_REQUEST_TIMEOUT: number = 300000; // 5 minutes in milliseconds
