--- src/external/services/ExternalMessageBroker/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ExternalMessageBroker';
