+++ src/external/services/LegacyClientMessageBroker/types/IInitOptions.ts
@@ -1,0 +1,8 @@
+// types
+import type { IBaseOptions } from '@common/types';
+
+interface IInitOptions extends IBaseOptions {
+  channel: BroadcastChannel;
+}
+
+export default IInitOptions;
