+++ src/external/services/LegacyClientMessageBroker/types/index.ts
@@ -1,0 +1,1 @@
+export type { default as IInitOptions } from './IInitOptions';
