+++ src/external/utils/createClientInformation/index.ts
@@ -1,0 +1,1 @@
+export { default } from './createClientInformation';
